+++ package.json
@@ -27,7 +27,6 @@
     "babel": "^6.23.0",
     "babel-loader": "^8.0.5",
     "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-env": "^1.7.0",
     "prop-types": "^15.7.2",
     "react": "^16.8.6",
